+++ javascripts/load_jquery.js
@@ -1,1 +1,0 @@
-window.$ = window.jQuery = require(__dirname+'/javascripts/jquery-2.2.2.min.js')