+++ CHANGELOG.md
@@ -5,4 +5,13 @@
 
 ### Goals
 * Integrate python calculation with website
-* Neural networks? Tensorflow?
+* Neural networks (sklearn 0.18)? Tensorflow?
+* ~~Add RidgeRegression~~
+* ~~Add EnsebleRegressors~~
+* Add Bayesian Regression
+* Add Lasso Regression
+* Add MPL Regression
+* Add Tensorflow Regression
+* Add animation when the calculator is processing data
+* Improve `comments` bar
+* Store calculation in `Number`, not in display